Update Read more link component to use icon link
