+++ source/03-components/_index.scss
@@ -33,7 +33,6 @@
 @use 'pager/pager';
 @use 'pager/pager--mini/pager--mini';
 @use 'progress/progress';
-@use 'readmore-link/readmore-link';
 @use 'site-logo/site-logo';
 @use 'site-name/site-name';
 @use 'site-slogan/site-slogan';
