+++ source/03-components/card/card.twig
@@ -38,7 +38,7 @@
         {% endif %}
         {% if read_more and url %}
           <div class="c-card__readmore">
-            {{ macros.readmore(url, title) }}
+            {{ macros.read_more(url, title) }}
           </div>
         {% endif %}
       </div>
