+++ source/03-components/read-more-link/read-more-link.scss
@@ -1,0 +1,7 @@
+// Component: Read More Link
+
+@use '00-config' as *;
+
+.c-read-more-link {
+  text-decoration: none;
+}
