+++ source/03-components/read-more-link/read-more-link.yml
@@ -1,6 +1,7 @@
 ---
-title: 'Title'
+modifier_classes: ''
 url: '#0'
+title: 'Title'
+label: ''
 description_prefix: ''
 hide_description: false
-modifier_classes: ''
