--- source/03-components/readmore-link/_readmore-link.scss
@@ -1,32 +1,0 @@
-// Component: Readmore Link
-
-@use '00-config' as *;
-@use '01-global' as *;
-
-.c-readmore-link {
-  padding-right: 28px;
-  text-decoration: none;
-
-  &:hover,
-  &:focus {
-    path {
-      fill: gesso-color(text, link-hover);
-    }
-  }
-}
-
-.c-readmore-link__icon {
-  display: inline-block;
-  margin-right: -28px;
-  position: relative;
-  top: 8px;
-  width: 28px;
-
-  path {
-    fill: gesso-color(text, link);
-  }
-}
-
-.c-readmore-link__accessibility-description {
-  @extend %visually-hidden;
-}
