--- source/03-components/readmore-link/readmore-link.stories.jsx
@@ -1,18 +1,0 @@
-import parse from 'html-react-parser';
-
-import twigTemplate from './readmore-link.twig';
-import data from './readmore-link.yml';
-
-const settings = {
-  title: 'Components/Read More Link',
-};
-
-const ReadMoreLink = args => (
-  parse(twigTemplate({
-    ...args,
-  }))
-);
-ReadMoreLink.args = { ...data };
-
-export default settings;
-export { ReadMoreLink };
