--- source/03-components/readmore-link/readmore-link.twig
@@ -1,22 +1,0 @@
-{% set classes = [
-  'c-readmore-link',
-  modifier_classes ? modifier_classes : '',
-]|join(' ')|trim %}
-
-{% spaceless %}
-  <a href="{{ url }}" {{ add_attributes({ 'class': classes }) }}>
-    {{ label|default('Read more') }}
-    <span class="c-readmore-link__icon">
-      <svg xmlns="http://www.w3.org/2000/svg" width="28" height="28" viewBox="0 0 28 28">
-        <path fill="#231f20" d="M14 4.648l9.352 9.352-9.352 9.352-1.641-1.641 6.508-6.563h-14.219v-2.297h14.219l-6.508-6.563z"></path>
-      </svg>
-    </span>
-
-    {% if not hide_description %}
-      <span class="c-readmore-link__accessibility-description">
-        {{ description_prefix|default('about') }}
-        {{ title }}
-      </span>
-    {% endif %}
-  </a>
-{% endspaceless %}
